Fix __callee__ for aliased methods
Contrary to backtraces, MRI uses the aliased name in  __callee__ and __method__.

Implementation-wise, it seemed too involved to change all callers of ThreadContext#pushCallFrame to take aliased names into account, so this changes the meaning of the name in DynamicMethod#call to reflect the name to use for __callee__ rather than what to use for backtrace.

In order to patch up backtrace generation, I changed (most) calls to ThreadContext.pushBacktrace to ignore the passed name and use the name from the method declaration instead. There are some cases in the IR Interpreter remaining, but those seem to either use a hardcoded `(eval)` or `(root)` or they extract a method from a Binding.

Bindings might need fixing up as well, but I'm not sure where that should be done.